Make splitwise payment collection multi-domain compatible (#2614)
This is kinda hacky because it works around the lack of multi-domain
stores. I opted for having the hacky parts only be used in the
multi-domain test for now. Should be easy to adjust to the stores
though once those are in place.

This does get us to the (exciting!) state where we can run a
multi-domain demo that starts on the splitwise domain, settles
payments on the public domain and transfers the balance update back!

This also includes a fix to the ACS through update stream stuff to
include transfer in/out as create/archives.

fixes #2353 fixes #2621