Hookup AcceptedAppPaymentTrigger to per-domain stores (#2667)
This changes the OnCreateTrigger to react to contracts on a given
domain and do lookups against that domain instead.

For the apps that already have domain configs, I set the right domain
everywhere. For the ones that don't, we go through
`defaultAcsDomain`. I've made that public for now but it will
obviously go away with #2638

fixes #2666