Introduce `ValidatorOnboardingConfig` (#2684)
Part of #2656

This config is not yet used for anything. The "secret" that is configured will later be required to match the secret in an existing `ValidatorOnboarding` contract (we use it for authenticating the validator candidate, i.e., for tying his party ID to his real-life identity).

Note than IRL, this config will likely be in a separate file from the main config that is only included if it exists at an expected location; i.e., `onboarding = include "onboarding.config"`. This `onboarding.config` is then what validator candidates get from their sponsoring SV via a secure "out of band" channel (like a Slack message :P).