Make treasury service load test more aggressive (#2681)
fixes #2373

I could not trigger the treasury service's backpressure even in more-aggressive load tests but I think we can close the issue nonetheless.
Even when the queue had up to ~85 events (note that I configured a buffer size of 40 for the test) no coin operations were being dropped/backpressured. I also spoke with @ray-roestenburg-da and he said that reliably triggering backpressure is not trivial - and so since I still didn't trigger it even with going up to 85 events in the queue I decided to not further spend time on trying to trigger backpressure/this DROPPED case for now. I also don't think (?) it's critical to test this code path because we are, e.g., already testing almost the exact same code path (this one, instead of this one) in the test and thus we know that the code behaves nicely when a StatusRuntimeException is thrown if the queue rejects an element.

Thus, all this PR does it make the treasury load test more aggressive.