Support wallet-user for app provider (#2691)
Supports auto-onboarding of a wallet user for an app provider. The service-user is now optional. If specified, and if different from the wallet-user, then another user is created for it, with the same primary party (will be useful for sharing a primary party between a login wallet user and a machine-to-machine token used by the backend).