Add cncluster reset to cluster workflow (#2742)
Happy to be told this is a bad idea for whatever reason but I had some
cases where I forgot to run this and thing went badly due to whatever
random state happened to be there before. And while I can ofc run it
myself, going for the option where people clean up after themselves
seems nicer to me.