Avoid confusing errors in CoinRulesRequest trigger (#2779)
* Avoid confusing errors in CoinRulesRequest trigger

This falls into the same readAs bug. I've changed the defaul to just
go for SubmitAndWait by default. That is also more efficient and in
all but one place we just ignore the transaction anyway.

For this trigger, the issue doesn't actually break anything. What happens if that the
submission fails but on retry we see that the contract is stale. But
the logs make no sense and look super confusing so not relying on this
working more by accident than anything else seems sensible.