Add TransferStore for in-flight transfers (#2725)
This adds a store which stores the transfers that are currently
in-flight. This will drive the TransferInTrigger implementation but
this PR does not yet switch it over (just for ease of review, I did
already test it).

This is slightly complex since we need to ingest into those stores
from multiple domains but overall it seems to work pretty well.

For now, this ingests through separate streams from ACS/transaction
tree. We could unify that but I'd like to wait until #2729 before we
try that.

fixes #2709