fix preflight bug (#2813)
The test was requiring that after login there is exactly one party ID seen on the page, but if a user was invited to a group, and logs in - there will be two. One in the header (the user's own party ID), and one in the invitation. In general, there are other scenarios where after the user logs in, they see party IDs on the page, so using "length(party-ID) should be 1" is incorrect.