Make `start-backends-with-local-frontend-testing.sh` more tested and robust (#2823)
Fixes #2408

### Makes `start-backends-with-local-frontend-testing.sh` more *tested*

We had a test for the bootstrap script that was used in the `sh` script but it wasn't creating the same environment as we get in the `sh` script. I fixed this, so bugs like #2480 and #2659 should be caught now. I also modify the test to run the bootstrap script twice in row.

### Makes `start-backends-with-local-frontend-testing.sh` more *robust*

Fixes its underlying bootstrap script to be able to run twice in a row, so that the `sh` script can be stopped and restarted without having to restart the long-running canton instance.