Canton bump - sprint 2306 (#2834)
fixes #2769

it also progresses #2467 by making it so that our diff-canton.sh script now correctly recognizes when we create a file in the forked Canton codebase. Thus, we no longer need to care in particular about not adding files in the forked Canton code base.
However, I am not closing #2467 as part of this PR as I didn't run this upgrade with the updated script end-to-end (I only noticed that this issue was still open as I was mostly done already) & there are likely still some improvements we want to make.