Restructure k8s cluster using namespaces (#2358)
This allows better isolation at the both the network and configuration manifest level. 

Note that the namespace name is included in the DNS name of the Kubernetes Service, so cross-namespace connection has been made explicit by including the namespace as part of the hostname.

Furthermore, secrets cannot be accessed from a different namespaces which keeps confidential data local to their use.