Migrate validator runbook + tests to new validator onboarding flow (#2874)
The validator runbook and the runbook preflight check are now based on the new onboarding flow it its entirety. We mock the "ask the SV operator to sponsor us" step by having validator candidates self-service themselves via the (still public) admin API of the SV app (the mid-term plan is to replace this with a dedicated client API endpoint a la "onboarding faucet").

We need to get an onboarding secret from an SV before we start the validator, because the onboarding secret needs to be put in the validator's config. This is not easy to do in the local runbook test, because here the SV apps and the validator app get configured at the same time. Instead, we use the "expected onboardings" variant of the flow here (which we originally implemented for initializing validators in our test environment.

Fixes #2836