Set nginx resolver everywhere and use fqdn (#2980)
We broke our cluster with a combination of two issues:

1. We started using variables in some placse in proxy_pass urls. That
changes nginx DNS resolution. The only fix seems to be to set the
resolver explicitly and use fully qualified domain names (fqdn)
2. We removed a seemingly unused case that still used the problematic
dns name (the wallet backend) without a variable. Apparently that one
meant that the new behavior thanks to the variable did not yet kick
in.

With the combination of both nginx just completely fails to resolve
the URL reproducibly. Adding this makes it work reproducibly (across 5
tries).

Rather than trying to be clever on where we need and don't need this I
just switched over all proxy_pass directives.

I could not find a way to dedup this nicely unfortunately so for now
this is duplicated. Once we autogenerate at least parts of that this
should look a lot nicer.

Given that I had to touch literally all of these directives I took the
liberty to also remove the proxy_http_version in all places that are
not grpc-web.

Tested through preflight on scratch.