Move java codegen methods out of LedgerApiAdministration (#2976)
* Move java codegen methods out of LedgerApiAdministration

It doesn't look like we'll be able to upstream that stuff soon (see
https://github.com/DACH-NY/the-real-canton-coin/issues/2914#issuecomment-1433189588)
so this PR takes a different approach and moves this into a file
outside of the the Canton fork.

I di have to change visibility of some methods. I won't try to
upstream that for now. I don't think this really causes problems when
bumping the fork and I'm still hoping that once the dust has settled
on the repo move we can upstream this.

fixes #2966

* fmt

* fix bootstrap

* .

* meh