Move ApiClientRequestLogger out of Canton fork (#2987)
Thereby reducing the size of our changes to the Canton fork.

We are not upstreaming the ApiClientRequestLogger, as Canton
does not really have a need for it: all the Ledger API client
calls typically originate from the console that also manages
the Ledger API server that receives these calls. So the server
logging is enough. Moreover, Canton has way less complex
interactions with the Ledger API than our apps.

Fixes #2915