Add documentation for running splitwell on separate domain (#2992)
* Add documentation for running splitwell on separate domain

fixes #2703

I ended up putting this as an additional section because it did get
fairly complex and in practice I also expect users will usually start
out on a single domain.

Also fixes some unfortunate issues in the existing docs. We considered
adding tests for this in the past but deprioritized this because we're
not investing in app dev experience. Unfortunately I think that’s
still the case so sticking with that :(