Switch splitwell invite flow to explicit disclosure (#3000)
* Switch splitwell invite flow to explicit disclosure

fixes #2898

fixes #1908 (by not doing that since we ended up going for explicit
disclosure after all)

This provides 3 main advantages:

1. We should no longer get a gradual slow down in the transactions
that create invites as more directory entries are created (which
happens by our normal preflight checks).
2. We speed up the already stupidly slow tests a bit by removing the
need to setup directory entries first. I kept them in one so we can
test that they're displayed correctly. This also applies to
runbooks. I tried to add a splitwell payment flow to some docs earlier
and gave up because the setup including directory entries was just too
many steps.
3. We exercise our explicit disclosure support for UIs. While we might
not need very much there for june, eventually we will and the Canton
team explicitly asked for usage feedback.

The UI for this is not particularly pretty but given that we're not
optimizing the splitwell UI, that seems fine.

* apparently you cannot end sentences in random places