Remove error check at the end of frontend tests (#3093)
fixes #1708

I kept it in the UI a discussed in #1708. It is still useful for users
until we do a better job at error handling in the frontend.

The checks here are particularly problematic since they can swallow
actual errors. E.g., we had a case where there was a scalatest
exception that caused the test to fail before it called
`consumeErrors` so then this piece here also failed and overwrote the
original exception.