Assert that payments are collected in round they are locked (#3096)
* Assert that payments are collected in round they are locked

fixes #1492

They already failed due to holding fees anyway. This just means that
now we have sensible error messages rather than random insufficient
founds.

Unfortunately this suffers a bit from `submitMustFail` not providing
any error message. We _could_ work around this using some form of
create and exercise and catching the exception server side. That
seemed like more complexity than it's worth though. In the end, we
care that it fails and the error message is more of a nice to have.