Cleanup scan transfer context APIs (#3112)
fixes #3214 and ensures that calls to get the coinrules/mining rounds now always use the caches. Previously, the getTransferContext call wouldn't use the ScanConnection caches & was confusingly sharing the same name as the similar, but slightly different, (App)TransferContext from the Daml code. I was able to completely remove the getTranferContext endpoints by instead implementing it as (cached) calls to getCoinRules and getOpenAndIssuingMiningRounds on the client-side.