Only instantiate http handlers once (#3167)
* Only instantiate http handlers once

fixes #3158

I abandoned my attempts to add a warning for this in the end. Tracking
`ledgerclient.connection` calls doesn't work. We create one per user
in the wallet so they can grow arbitrarily large. Tracking the number
of created handlers in a static variable doesn't work because we spin
up one per app.

What would work is tracking how many handlers we create per app. But
that seems to require an excessive amount of infrastructure for this.