Add multiple splitwell domain configs (#3182)
* Add multiple splitwell domain configs

fixes #3057

Only adds it to the config and exposes it on the endpoint. This does
not yet use the domain anywhere.

I went for a list of other domains rather than just two. While two
should be enough for now, keeping the list is just as easy and seems
to make the code a bit less ad-hoc.