Bump resource class to 2xlarge+ (#3201)
We are out of memory which makes tests fail left and right which is
not great obviously. This is the highest resource class though (for
docker, we could go to a machine executor to get a bit more ram (and
confusingly fewer cpus)).

I'll see if I can forcibly reduce the JVM memory of Canton after CI
stops failing. We may be somewhat out of luck there though. Our
topology has grown a lot.