Hack around missing synchronization in update service (#3195)
fixes #3192

This isn't a very principled fix because this really shouldn't be
necessary in the first place. See #2728. But we've seen it at least
twice now so I don't just want to ignore it.

In theory tons of commands can run into this but trying to slap
retries for this everywhere seems like a lost cause so let's add them
where we actually encounter it.