Move directory app to svc participant (#3210)
* Move directory app to svc participant

This is what we do in the cluster already and as soon as we do
decentralization this will become inevitable.

I've moved the upload into the sv app. That seems consistent to what
we do for the core coin dar and makes sense given that we expect to
merge the directory logic into the sv app eventually.

Tested preflight on scratch.