Bump splitwell accept timeout (#3243)
* Bump splitwell accept timeout

We've seen this fail once on staging. 5s sounds like a lot but with
multi-domain. This includes a transfer out + transfer in now so it's
not super surprising that this needs some extra time.

I did look at the logs and can see that it succeeds just after the 5s.

fixes #3242