rename method in wallet frontend, and remove redundant argument (#3253)
Apparently, we have in the backend `onboard` as an admin action to onboard an arbitrary user and `register` to onboard oneself. We're actually using the latter from the UI (as we should), but internally still calling it `onboard` in the code. To make things even more confusing, we're still passing the user ID to that method, which then goes unused.