Enable oidc logging (#3270)
For now just hardcoded to debug until we sort out the issue. It is
pretty verbose but only during the login step so in the overall test
logs it's still barely visible. Added a TODO though, I expect we'll
want to at least lower this back to info later on.

Hopefully gets us closer to figuring out #3245.