Support splitwell domain disconnects (#3277)
* Support splitwell domain disconnects

fixes #3247

This is only one part of domain disconnects #2210 covers handling this
everywhere.

To make this work properly I switched to starting the triggers on all
domains. This is fine from correctness PoV but not great for
performance. I opened #3285 for that but will likely only tackle that
later. For now I care more about making it work than about making it
fast and with domain migrations the domains those triggers run on will
change anyway.