Make shared frontend tests share the frontend (#3400)
* Make shared frontend tests share the frontend

You might guess that a test called
`FrontendIntegrationTestWithSharedEnvironment` would also share the
frontends across tests.

However that is not the case. I fixed that and added some more logging
to debug #3398 and similar issues easily.

Does this actually make a difference? Yes, and a very significant
one. `WalletPaymentFrontendIntegrationTest` takes 150s for me locally
without this and 80s with this. That's not quite a 2x improvement but
pretty close.

The downside is that we have to clean up at the end of the test but
that seems like a price worth paying.