Merge per-domain stores into one (#3366)
fixes #3363

This avoids two main issues:

1. We don't need to worry whether the target store is up during
ingestion. Currently, we can in theory deadlock when we see a transfer
out on one domain but the target domain is not connected because we
wait for that store to appear.
2. We only track a single offset for a given domain (well atm still
two if you include the acs but we'll fix that). That is important
so we actually know where to resume.

Because we track everything by transfer id (source, transfer out id),
this requires very little changes.

The automation is still per domain. This matters because we cannot
submit commands if the domain is not connected.