Synchronize before logging out (#3445)
* Synchronize before logging out

Based on the logs that I've stared at it looks like sometimes we call
logout while still on the original page, e.g., splitwell. We then get
redirected to the wallet and accept it. Once we get redirected back
we're now no longer logged in.

I've also added some logs to the frontend to make sure it's clearer
that a logout is happening.

fixes #3439

maybe also related to #3431 but haven't looked into it in detail.