Improve documentation and tooling around management of dns secrets (#3423)
* Improve documentation and tooling around management of dns secrets

We ran into some issues where we accidentally seem to have deleted
secrets for clusters that were still in use. To make things more
confusing this only fails once the tls cert needs to be
renewed (either because it expires or because the domain names
change).

This PR adds some more tooling and docs around this to hopefully make
this clearer in the future.

I manually updated all secrets so atm we have 7 secrets one for each
cluster.