Increase timeouts in WalletSurviveCantonRestartIntegrationTest.scala
fixes #3584. Looking at the logs was a bit confusing at first. The problem is that the tap(2) timed out. The reason why the logs were confusing is that before shutdown succeeded, it actually completes, which made it look like it succeeded. Thus, just increasing the timeout here (Canton is slow to start) should be enough.