Add multi-domain ACS store (#3582)
* Add multi-domain ACS store

This adds a multi-domain ACS store that tracks contracts across all
domains. This is pretty much required to do efficient reads if we have
domain upgrades.

The eventual goal is that this completely replaces the existing
stores. However, to ease review this PR only does the bare minimum and
adds listContracts and lookupContractById. However, the infrastructure
should largely be sufficient.

The remaining work is tracked in #3181

The unit tests should be relatively comprehensive.

Logging is lagging, I started adding IngestionSummary-style logging
but it just made things harder to review so I ripped it out
again. Adding that will be my next step before hooking it up anywhere.