Raise `requiredNumConfirmations` to 2/3  (#3579)
Matching the threshold used for voting-based decisions.

Fixes #3558

Note that with this PR, we will also start enforcing that all of our SVs approve the onboarding of all our other SVs (except that of SV1, which as a founding SV doesn't need approval). This makes sense semantically and may also make SVC bootstrapping slightly more robust and easier to debug.