Add equivalent of streamContracts to multi domain acs store (#3675)
This one is slightly different since the same contract can appear
multiple times. That will be required to make our automation tick
properly. E.g., consider automation that only wants to react to
contracts on a given domain.

The event numbers on contractLocations didn't quite work out as I was
hoping for. The problem is the case where we see the transfer in
before the transfer out. The transfer out will only remove an event
but not add a new one so a stream that just keeps reading events will
never get notified. Instead I switched to only tracking the single
state per contract and bumping the counter for that.

I kept the logging based on Contract locations. Generally we try to
keep our logging to only log diffs so logging individual domains
added/removed seems sensible.

part of #3181