Make TxLogStore APIs multi-domain compatible (#3717)
The APIs are close enough to what we can make multi-domain compatible
that rather than duplicating everything I've opted for just slightly
extending them.

For now this does per-domain reads. Until we do global domain upgrades
that should be good enough. But I think we can probably get rid of it
for the wallet usecase fairly easily by paginating based on synthetic
sequential event ids rather offsets. But not in this PR.

part of #3181