Download preflight cluster logs to CI artifacts (#3722)
Fixes #3572

This makes our CI download logs from a selection of apps/pods and make them available via the "Artifacts" tab for preflight check runs.

I've hardcoded the list of apps. Ideally this list should be generated programmatically out of our deployment description, but I didn't even want to start with this given that the way we are defining our deployments is going to change soon. But fear not: should the list become outdated the step I'm adding here should still not cause CI to "fail"; worst case you'll just not get any useful logs.

In order for the log downloading to work, I have added the "Private Logs Viewer" role to cirleci via the google cloud IAM (for both da-cn-testnet and da-cn-scratchnet).