Switch ingestion to poll based on ledger end offset (#4019)
* Switch ingestion to poll based on ledger end offset

fixes #3921

This switches the ingestion to stream until the current ledger end and
advance the offset in the store at the end. This guarantees that as
long as we see events on any domain, the offset on all domains will
eventually advance which means that the minimum that we use for
command deduplication will never get stuck. See #3921 for a more
detailed description.