Improve naming of SV onboarding-related things (#4226)
Fixes #3378

List of naming changes; each line roughly corresponds to a commit on the PR:

- `onboardSv` -> `startSvOnboarding`
- `/onboard/sv` -> `/onboard/sv/start`
- `SvOboarding` -> `SvOnboardingRequest`
- `SvConfirmed` -> `SvOnboardingConfirmed`
- `SvcRules_ConfirmSv` -> `SvcRules_ConfirmSvOnboarding`
- the onboarding state called `onboarding` (this is where confirmations are collected) -> the onboarding state called `requested`
- `SvBootstrapConfig` -> `SvOnboardingConfig`
- `bootstrapping` -> `onboarding` (in docs and runbook)
- `bootstrapType` -> `onboardingType` (pulumi)

The rationale between the "bootstrapping" refactor is that we only want to use bootstrapping when referring to "network bootstrapping" or things where we can't change the naming - like Canton's bootstrapping stuff. In all other relevant cases we want to use "onboarding" and/or "initialization".