Invalidate circleci class cache on all openapi schema changes (#4247)
* Invalidate circleci class cache on all openapi schema changes

Apparently guardrail’s SBT definitions are messed up even if you don't
do imports so this PR instead just nukes the whole cache on any
openapi schema change. We don't change the schemas that often so
while this makes me very sad, I don't expect it to be a big issue in practice.