Include validator app in SV runbook (#4380)
* Include validator app in SV runbook

part of #4319

This PR adds the validator app to the SV runbook and briefly explains
how you can deploy the wallet UI against it. I didn't go into two much
detail there since it's exactly the same process as for the
self-hosted runbook so some cross references seemed more appropriate.

This does not yet touch the helm setup mainly because i am scared of that.