Extend helm runbook with instructions for deployment of SV wallet UI (#4402)
* Extend helm runbook with instructions for deployment of SV wallet UI

part of #4319

Got it to work in the end after all. I don't think any of our
customers will actually manage to setup the ingress correctly in the current
setup but at least the docker image is fine now and we are unblocked
for our own testing (provided we tackle #4387 because atm probably
only Mike and I know how to test this stuff).