Synchronize package upload between validator and sv app (#4706)
* Synchronize package upload between validator and sv app

Canton topology management is buggy and falls apart if uploads happen
concurrently. Given that we wait for the validator app to do its thing
anyway waiting for the package upload as well shouldn't meaningfully
slow things down.

The Canton team will make sure that's fixed in the new BFT topology
management but not try to fix the old one for now. Canton issue is
tracked in https://github.com/DACH-NY/canton/issues/12993

fixes #4685