Move service cluster metadata fields into label (#4760)
I got a ton of errors on scratchnet about this. My best guess is that
we upgraded to a new k8s version on scratchnet (but confusingly not
yet on devnet and maybe other clusters) that doesn't allow for this. I
don't think it was what we intended either so this change seems reasonable.