Fix devnet deploy jobs (#4759)
Currently failing because it expects a boolean but gets a string. I
can't easily test this but the fix seems fairly sensible. Will take a
look tomorrow to confirm that it works.

The release candidate job on the other hand works but did not tag
things as a release candidate because it defaults to false …

However, because we manually triggered one of those to get the docs
helm chart removal in, the tag does exist and is correct so at least
for this week we don't need to do anything to fix that.