Use the same port for all SV apps (#4931)
* Use the same port for all SV apps

And stop exposing them externally. We have nice URL for that.

Tested the pulumi runbook and the jsonnet deployment.

This should also fix the pulumi cluster deployment which is currently
just completely broken for sv2-sv4. Haven't tested that in detail just
yet though because a bunch of other stuff failed. Will do that next.

fixes #4641