Ensure that SV1-4 have voted for a coin price in cluster (#4969)
By doing so, single SVs can't modify the coin price to absurdly high values, which would break the network.

We could also have set the initial coin price vote via the Daml flow for adding a member, but the solution I chose here (doing an update from out of the SvApp init, after onboarding is done) seems less invasive. Happy to reconsider this in the future. (But we should get this merged quickly + it's certainly good enough for testnet.)

Fixes #4862