Fix synchronization in splitwell frontend upgrade tests (#5084)
fixes #5082

This also fixes the confusing initialization failures in the auth0
tests. What happens is that when this error happens, the participants
seems to get bricked and later on GetConnectedDomains no longer
returns that party on the domain we submitted the transfer out on.

I've exceeded my quota for the number of Canton issues I'm allowed to
report this week but will reach out to them next week.