Set `joinedAsOfRound` correctly for new SV members (#5075)
How we do it:

We require our SV candidates to show us the three active open mining rounds. It has to be all three of them, so they can't just use the oldest one and claim it's the newest one, thereby getting rewards for rounds they weren't (as) involved in. We then take the newest mining round to calculate the next mining round number - and this is then the `joinedAsOfRound` and the first round they will receive rewards for.

Fixes #4749