Move distributed domains test to a separate CI job (#5223)
Keeping them in the same job is going to be problematic because the
sequencer/mediator connections persist by default so we'll copy the
fragility of the distributed domain setup to the non-distributed
one. Hopefully we can get rid of that at some point but for now this
at least allows us to iterate on both in a reasonable state.

fixes #5220

I opened #5201 to think more carefully about our test setup
there. Input welcome.