Reenable test for distributed domain (#5238)
* Reenable test for distributed domain

fixes #5197

According to Gerolf, today’s snapshot turned on mediator group
addressing which means that topology requests end up with the right
mediator.

This is turned on automatically (and cannot be turned off) so nothing
for us to actually change.

The protobuf APIs changed in incompatible ways. There isn't yet a new
Canton OSS dump with the new changes so for now I somewhat surgically
copy pasted the changes (trying to do it in a non-surgical way did not
go well)

Still waiting for confirmation for whether ignoring the error makes
sense. The tests seem to pass though. Worst case I'll disable it again.