Upgrade Canton and remove party allocation workaround (#5348)
* Upgrade Canton and remove party allocation workaround

fixes #5300

The retries in the current definitely don't make sense. If anything we
should retry with the same hint to keep things idempotent. But we only
allocate in user requests where the client retries and during init
where if it fails our app gets restarted and we retry the whole init
so I don't think we want a builtin retry here.