Migrate `DuplicateTransferTest` to XNodes (#5310)
Fixes #5257

The "create contract" and "transfer out" parts of this test were flaking (something something parties not known on all domains). This is not the main thing we are testing here there and is likely related to initializing the test, so I added an eventually, which stops the test from flaking.