Avoid flakiness in unexpected tx log events tests (#5420)
fixes #5386

Seems like the simplest option for now. If we end up with tons of tests for
the SV wallet we may want to add an option to disable the automation
for that or write a helper that ignores all reward collection events
but for one case that doesn't seem worth doing.