Make default app init optional and use custom init for SV app (#5585)
As part of making the SV app bootstrap the domain, the default init
falls apart because it assumes that we are already connected to a
domain. Trying to add enough overrides to make it work made things
more confusing than helpful so this PR instead splits CNNode into
CNNodeBase which does the mandatory init all apps need and CNNode
which provides the default "simple" initialization logic used by all
apps but the SV app.

Also adds a ensureThatWithResult variant.