Make scan in cluster use sv1's user instead of the svc user (#5538)
Fixes #5507

Also removes code that allocates a user for scan in the cluster. We don't need a user for scan (anymore).

I tried reusing the new Helm chart for scan but ran into problems and decided to time-box this.
This might be more fun once we deprecate the svc namespace and/or the jsonnet-based deployment route.