Initialize domain and SVC party from SV app (#5577)
fixes #5478
fixes #5477

This is currently only activated when testing together with a
distributed domain. I think it makes more sense to enable it in all
cases evevn for a centralized but I'd like to wait until we've at least fixed the domain
connections issues and ideally completed the switch to X nodes and
killed the SVC app.