Set serials for all topology requests and add retries (#5629)
* Set serials for all topology requests and add retries

As explained in #5616 not setting serials is a trap. Our logic is
based on reading an existing state and then setting it to an updated
state. Without setting serials that can just overwrite intermediate
updates.

I'm trying to lobby the Canton team to make serials mandatory since
I'm not convinced you can reasonably use the current API without
setting serials but that's somewhat orthogonal.

The retries are horrible. I've opened #5628 to improve that. Gerolf
said he'll at least provide a string we can match on.

fixes #5616
fixes #5609