Add SV participants to unionspace (#5637)
* Add SV participants to unionspace

fixes #5480

Seemed easiest to do this as part of the same request as the
PartyToParticipantX mapping. We'll definitely need to revisit this
once we add proper reconciliation loops but for now I don't see any
value in making it more complex.

Given that the whole validation logic in Canton is missing this
doesn't actually enable anything particularly useful atm afaict.