Fix scripts for local frontend testing for X nodes (#5780)
fixes #5778

Along the way removes the old unused topology files because I'll break
them by accident anyway.

I've also removed the minimal bootstrap file. It would make sense with
a minimal Canton topology but without that it just seems like a waste
of effort. The resources are used up by the number of nodes not by the
startup script.