Add a global lock around node onboarding and topology transactions (#5847)
fixes #5856

slaps roof of CN: I can fit so much locking in this bad boy

This is absolutely terrible but it is very effective. I haven't
managed to get it to fail with this across ~20 CI runs and it even
seems to work for the simtime tests (will switch those over in a
separate PR).

It does slow down things but not that bad, ~10% at worst in CI
times. Often not noticeable at all.

I did have to bump the retries in the RetryProvider. Otherwise, I
occasionally see the validator exceeding its retries for the getUser
call. There are opportunities for being more clever with locks here
but tbh I don't care enough given that the overall CI slowdown seems
reasonable and this is temporary. We can improve it later if it does
turn out to be a problem.

This does only work if SVs and validators have onboarding configs
specified. This is the case in basically all our tests and in our
runbooks. I didn't also want to mess around with config changes in
this PR. We could either just revert the changes to make it optional
again (you know my opinion on that …) or change the config such that
only the SV connection is mandatory which is probably cleaner but also
seems like wasted work for a workaround. Or we do nothing and just
rely on noone using it.

We do need to add lock expiry but I'd suggest to do that as a follow-up.