Ingest from participant begin in Scan time integration test (#6019)
fixes #5880

This is actually something that we may want to enable for SV 1 even in
prod eventually since we likely want to guarantee that we observe the
initial bootstrapping.

But for now, only enabling this in the one test that flaked a lot. I
figure we an easily enable it as needed in other tests but I don't
have time to figure out which ones atm.

The string conversion stuff is kinda crappy but seemed better than
trying to rebuild the whole multi domain store API for this.