Fix deployment sequencing (#6108)
fixes #6075

There are zero docs on the meaning of those statuses so I'm somewhat
guessing but it seems very plausible that failing means something in
the workflow failed but parts are still running while "failed" means
it has failed completely.

I don't know how to tets this unfortunately and we don't output enough
info to see the output that caused problems. I've changed it to output
the response even in the case where we do continue which should
hopefully help.