Reduce pulumi dependencies of SV participant (#6114)
It does not actually depend on the domain and most of the
secrets (the connection to the domain is established by the SV app
which depends on both). Given that both the participant and the domain are some of
the slowest pods to startupt this has a pretty significant impact of
making deployment almost a minute faster.

Unfortunately it's still not as fast as I would expect it to be. There
is a very noticeable pause after postgres and the secrets are deployed
before it deploys the participant and the domain. I don't really have
a better answer than blaming it on the pulumi scheduler at this point.