Restrict approved SV identities on non-devnet (#6156)
fixes #6147

Note that on dev-net the setup is now slightly different: Only sv1-4
aprove the sketchy SVs. So if we have enough external SVs at some
point we'll struggle. That seems fine to me. We may eventually ask
them to approve more if we setup a real devnet.

This also has the consequence that the SV runbook does not work by
default. If we need to run it during testing we can easily use the
endpoint to approve it.