Write ACS dump to GCP bucket (#6191)
* Write ACS dump to GCP bucket

part of #6146

This currently still writes to the testing bucket Simon setup. I
expect we will want to keep a separate bucket for integration tests
and the actual cluster so for now I'll just stick with the hacky one
in integration tests and will make sure we have a cleaner one in the
cluster.

I kept the mode to dump to a directory for now, might remove it later
but it's cheap to keep around and may be useful for testing and debugging.