Wire up ACS exports in non-devnet clusters (#6203)
part of #6146

Also adds a basic test. We could also run this in devnet clusters in
theory but I don't want to worry about having to prune that bucket for
now.

This still writes to the scratch bucket with no sensible directory
structure. Next PR will add a new bucket and a better file structure.