Bump Canton to get init_id for X nodes (#6173)
fixes #6238
fixes #6031

We probably want to bump the ACS pruning interval to make this less
likely. However, I would like to make sure that our logic is actually
robust and that would just hide issues so for now I'm not doing that
and instead opened #6261 for that. Will do that next week or so if it
doesn't break before then.